server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aerich[toml]>=0.8.1",
    "aiohttp>=3.11.13",
    "bcrypt>=4.2.1",
    "fal-client>=0.5.9",
    "fastapi-mail>=1.4.2",
    "fastapi[standard]>=0.115.8",
    "google-auth>=2.27.0",
    "openai>=1.12.0",
    "pydantic-settings>=2.7.1",
    "python-dotenv>=1.0.1",
    "python-jose[cryptography]>=3.4.0",
    "replicate>=1.0.4",
    "requests>=2.32.3",
    "tortoise-orm[asyncpg]>=0.24.0",
    "asyncpg>=0.29.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
]

[tool.aerich]
tortoise_orm = "src.config.settings.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
